django projects/django shop blog social/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-mdeditor==0.1.20
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
factory-boy==3.2.1
Faker==14.1.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.14.0
oauthlib==3.2.0
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
pycodestyle==2.9.0
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
psycopg2==2.9.3
gunicorn==20.1.0